Var latex

+++ cosmos/run.py
@@ -31,9 +31,10 @@ parser.add_argument("pdfdir", type=str, help="Path to directory of PDFs")
 parser.add_argument('-d', "--weightsdir", default='weights', type=str, help="Path to weights dir")
 parser.add_argument('-w', "--weights", type=str, help='Path to weights file', required=True)
 parser.add_argument('-t', "--threads", default=160, type=int, help="Number of threads to use")
-parser.add_argument('-n', "--noingest", help="Ingest html documents and create postgres database")
+parser.add_argument('-n', "--noingest", help="Ingest html documents and create postgres database", action='store_true')
 parser.add_argument('-o', "--output", default='./', help="Output directory")
 parser.add_argument('-p', "--tmp_path", default='tmp', help="Path to directory for temporary files")
+parser.add_argument('--debug', help="Ingest html documents and create postgres database", action='store_true')
 
 args = parser.parse_args()
 
@@ -68,11 +69,14 @@ def preprocess_pngs(img_f):
 
 def convert_to_html(xml_f):
     xpath = os.path.join(xml, xml_f)
+    print('Begin convert')
+    print(xpath)
     l = xml2list(xpath)
+    print(l)
     list2html(l, f'{xml_f[:-4]}.png', img_d, html)
 
 def match_proposal(proposal_f):
-    proposal_f_full = os.path.join(f'{tmp}', 'cc_proposals', proposal_f)
+    proposal_f_full = os.path.join(f'{tmp}', proposal_f)
     xml_f = f'{xml}/{proposal_f[:-4]}' + '.xml'
     process_doc(xml_f, proposal_f_full, xml_f)
 #for pdf_f in os.listdir(args.pdfdir):
@@ -106,7 +110,6 @@ results = [pool.apply_async(write_proposals, args=(os.path.join(f'{tmp}', 'image
 print('Begin preprocessing pngs')
 results = [pool.apply_async(preprocess_pngs, args=(x,)) for x in os.listdir(os.path.join(f'{tmp}', 'images'))]
 [r.get() for r in results]
-shutil.rmtree(f'{tmp}/images2')
 
 with open('test.txt', 'w') as wf:
     for f in os.listdir(f'{tmp}/images'):
@@ -160,13 +163,22 @@ for idx, image_id in enumerate(tqdm(image_ids)):
     zipped = zip(r["class_ids"], r["rois"])
     model2xml(info["str_id"], xml, [1920, 1920], zipped, data_test.class_names, r['scores'])
 
-results = [pool.apply_async(match_proposal, args=(x,)) for x in os.listdir(os.path.join(f'{tmp}', 'cc_proposals'))]
+results = [pool.apply_async(match_proposal, args=(x,)) for x in os.listdir(f'{tmp}') if x[-4:] == '.csv']
 [r.get() for r in results]
 
-for xml_f in os.listdir(xml):
-    xpath = os.path.join(xml, xml_f)
-    l = xml2list(xpath)
-    list2html(l, f'{xml_f[:-4]}.png', f'{tmp}/images', html)
+if not os.path.exists(html):
+    os.makedirs(html)
+    os.makedirs(os.path.join(html, 'img'))
+if not os.path.exists(os.path.join(html[:-4], 'latex')):
+    os.makedirs(os.path.join(html[:-4], 'latex'))
+print('Begin converting to html')
+results = [pool.apply_async(convert_to_html, args=(x,)) for x in os.listdir(xml)]
+[r.get() for r in results]
+print('End converting to html')
+#for xml_f in os.listdir(xml):
+#    xpath = os.path.join(xml, xml_f)
+#    l = xml2list(xpath)
+#    list2html(l, f'{xml_f[:-4]}.png', f'{tmp}/images', html)
 
 # postprocess
 tmp_html = html.replace('html', 'html_tmp')
@@ -177,8 +189,9 @@ print("Running postprocessing")
 postprocess.postprocess(html, tmp_html)
 
 # replace old html with corrected stuff.
-shutil.rmtree(html)
-shutil.move(tmp_html, html)
+if not args.debug:
+    shutil.rmtree(html)
+    shutil.move(tmp_html, html)
 
 # Parse html files to postgres db
 input_folder = ingestion_settings['input_folder']
@@ -194,11 +207,8 @@ db_connect_str = ingestion_settings['db_connect_str']
 strip_tags = ingestion_settings['strip_tags']
 ignored_file_when_link = ingestion_settings['ignored_file_when_link']
 
-if args.noingest:
-    parse_html_to_postgres(input_folder, output_html, merge_folder, output_words, output_equations, db_connect_str, strip_tags, ignored_file_when_link, store_into_postgres=False)
-else:
+if not args.noingest:
     parse_html_to_postgres(input_folder, output_html, merge_folder, output_words, output_equations, db_connect_str, strip_tags, ignored_file_when_link, store_into_postgres=True)
 
-
-#shutil.rmtree('xml')
-shutil.rmtree(f'{tmp}')
+if not args.debug:
+    shutil.rmtree(f'{tmp}')
